Refactoring: - separate user.hpp and common.hpp into subfiles. - cpp/hpp files use now four spaces indentation
